\name{conditionalStatements}
\alias{conditionalStatements}
\title{conditional Statements used in R
      }
\description{this function shows ho to use conditional Statements
}
\usage{
conditionalStatements()
}
\arguments{
\item{type}{data }
}

\examples{
conditionalStatements()

}
\author{ZekaiOtles}