\name{createDataFile}
\alias{createDataFile}
\title{Create Data File
      }
\description{this function reada the file or ask for files  in R
}
\usage{
createDataFile(type=type)
}
\arguments{
\item{type}{data }
}

\examples{
%createDataFile(fname="/home/otles/NetBeansProjects/uRProgramming/data/07Haziran2010_sariyer.csv")
createDataFile()
}
\author{ZekaiOtles}