\name{matrixDataStructures}
\alias{matrixDataStructures}
\title{matrixData structures used in R
      }
\description{this function shows ho to use matrixData structures
}
\usage{
matrixDataStructures()
}
\arguments{
\item{type}{data }
}

\examples{
matrixDataStructures()

}
\author{ZekaiOtles}